Move helper classes to _internals submodule
